+++ src/deadline_timer.hpp
@@ -24,7 +24,7 @@ using reset_atom = atom_constant<atom("reset")>;
 using deadline_timer = typed_actor<reacts_to<reset_atom>>;
 
 deadline_timer::behavior_type
-deadline_timer_impl(deadline_timer::pointer self, int timeout);
+deadline_timer_impl(deadline_timer::pointer self, size_t timeout);
 
 } }
