+++ src/gate_service.hpp
@@ -56,7 +56,7 @@ private:
 
 gate_service::behavior_type
 gate_service_impl(gate_service::stateful_broker_pointer<gate_service_state> self,
-                  int timeout, const std::string& log);
+                  size_t timeout, const std::string& log);
 
 } }
