+++ src/socks5_service.hpp
@@ -58,7 +58,7 @@ private:
 
 socks5_service::behavior_type
 socks5_service_impl(socks5_service::stateful_broker_pointer<socks5_service_state> self,
-                    int timeout, bool verbose, const std::string& log);
+                    size_t timeout, bool verbose, const std::string& log);
 
 } }
